#20 refactor: 녹음 기능 수정
